# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriScan AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Vitamin {
  name: string;
  amount: string;
  dailyPct: number;
}

export interface Mineral {
  name: string;
  amount: string;
}

export type RecommendationType = typeof RecommendationType[keyof typeof RecommendationType];


export const RecommendationType = {
  positive: 'positive',
  warning: 'warning',
  tip: 'tip',
} as const;

export interface Recommendation {
  emoji: string;
  text: string;
  type: RecommendationType;
}

export interface ScanResult {
  id: number;
  foodName: string;
  confidence: number;
  servingSize: string;
  calories: number;
  protein: string;
  carbs: string;
  fat: string;
  fiber: string;
  sugar: string;
  sodium: string;
  cholesterol: string;
  healthScore: number;
  healthSummary: string;
  vitamins: Vitamin[];
  minerals: Mineral[];
  allergens: string[];
  dietaryFlags: string[];
  ingredients: string[];
  recommendations: Recommendation[];
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export type NutritionStatsTopFoodsItem = {
  foodName: string;
  count: number;
};

export type NutritionStatsAvgMacros = {
  protein: number;
  carbs: number;
  fat: number;
};

export interface NutritionStats {
  totalScans: number;
  avgCalories: number;
  avgHealthScore: number;
  topFoods: NutritionStatsTopFoodsItem[];
  avgMacros: NutritionStatsAvgMacros;
}