# lib/api-zod/src/generated/types/mineral.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriScan AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Mineral {
  name: string;
  amount: string;
}