# lib/api-zod/src/generated/types/nutritionStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriScan AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NutritionStatsAvgMacros } from './nutritionStatsAvgMacros';
import type { NutritionStatsTopFoodsItem } from './nutritionStatsTopFoodsItem';

export interface NutritionStats {
  totalScans: number;
  avgCalories: number;
  avgHealthScore: number;
  topFoods: NutritionStatsTopFoodsItem[];
  avgMacros: NutritionStatsAvgMacros;
}