# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriScan AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecommendationType } from './recommendationType';

export interface Recommendation {
  emoji: string;
  text: string;
  type: RecommendationType;
}