# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriScan AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Mineral } from './mineral';
import type { Recommendation } from './recommendation';
import type { Vitamin } from './vitamin';

export interface ScanResult {
  id: number;
  foodName: string;
  confidence: number;
  servingSize: string;
  calories: number;
  protein: string;
  carbs: string;
  fat: string;
  fiber: string;
  sugar: string;
  sodium: string;
  cholesterol: string;
  healthScore: number;
  healthSummary: string;
  vitamins: Vitamin[];
  minerals: Mineral[];
  allergens: string[];
  dietaryFlags: string[];
  ingredients: string[];
  recommendations: Recommendation[];
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}