# lib/api-zod/src/generated/types/vitamin.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriScan AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Vitamin {
  name: string;
  amount: string;
  dailyPct: number;
}